+++ package.json
@@ -43,7 +43,7 @@
     },
     "scripts": {
         "start": "react-scripts start",
-        "build": "rm -rf dist && NODE_ENV=production babel --extensions .ts,.tsx src/lib --out-dir dist --copy-files && tsc",
+        "build": "rm -rf dist && NODE_ENV=production babel --extensions .ts,.tsx src --out-dir dist --copy-files && tsc",
         "test": "react-scripts test",
         "eject": "react-scripts eject"
     },
