--- src/lib/index.tsx
@@ -1,3 +1,0 @@
-import { SearchBar as TurboSearchBox } from "./components/search/search-bar";
-
-export { TurboSearchBox as default };
